Prefer to use std::shared_ptr instead of std::tr1::shared_ptr etc
This patch replaces uses of TR1 utilities such as std::tr1::shared_ptr
and std::tr1::bind with the C++11 equivalents (when available).

The solution used is to define a new namespace alias, stdext, and then
refer to stdext::shared_ptr instead of std::tr1::shared_ptr, and
similarly for weak_ptr and bind. When neither C++11 or TR1 is
available, Boost is used as before, and the Boost versions of those
utilities are added to namespace stdext.

A similar change was already made to prefer std::bind to std::tr1::bind,
in commit 79b7394db7623f6d588b048fca1ab6f2266731f8. That change to
tasks.cpp can be simplified by reusing the stdext alias introduced by
this patch.

Fixes:
https://projects.vdr-developer.org/issues/2611
https://projects.vdr-developer.org/issues/2612